src/models/stable_diffusion_2_txt2img_512/requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.13.1+cpu
optimum-graphcore
pillow